fix: close final PR review findings
